CSCTTV-4140 TTV model update

+++ aspnetcore/src/DatabaseContext/Entities/FactFieldValue.cs
@@ -9,7 +9,7 @@ public partial class FactFieldValue
 
     public int DimFieldDisplaySettingsId { get; set; }
 
-    public int DimNameId { get; set; }
+    public long DimNameId { get; set; }
 
     public int DimWebLinkId { get; set; }
